+++ src/main/java/com/Sucat/domain/user/exception/UserException.java
@@ -1,6 +1,6 @@
 package com.Sucat.domain.user.exception;
 
-import com.Sucat.global.error.ErrorCode;
+import com.Sucat.global.common.code.ErrorCode;
 import com.Sucat.global.error.exception.BusinessException;
 
 public class UserException extends BusinessException {
