+++ src/main/java/com/Sucat/global/common/code/BaseCode.java
@@ -1,6 +1,6 @@
 package com.Sucat.global.common.code;
 
-import com.Sucat.global.dto.ReasonDto;
+import com.Sucat.global.common.dto.ReasonDto;
 
 public interface BaseCode {
     public ReasonDto getReason();
