+++ src/main/java/com/Sucat/global/common/constant/SwaggerConstants.java
@@ -1,4 +1,4 @@
-package com.Sucat.global.util;
+package com.Sucat.global.common.constant;
 
 public class SwaggerConstants {
     /**
