+++ src/main/java/com/Sucat/global/common/dto/ErrorReasonDto.java
@@ -1,0 +1,14 @@
+package com.Sucat.global.common.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ErrorReasonDto {
+    private HttpStatus httpStatus;
+    private final boolean isSuccess;
+    private final String code;
+    private final String message;
+}
