+++ src/main/java/com/Sucat/global/common/dto/ReasonDto.java
@@ -1,4 +1,4 @@
-package com.Sucat.global.dto;
+package com.Sucat.global.common.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @Builder
-public class ErrorReasonDto {
+public class ReasonDto {
     private HttpStatus httpStatus;
     private final boolean isSuccess;
     private final String code;
