+++ src/main/java/com/Sucat/global/config/SwaggerConfig.java
@@ -8,7 +8,7 @@ import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import static com.Sucat.global.util.SwaggerConstants.*;
+import static com.Sucat.global.common.constant.SwaggerConstants.*;
 
 
 @OpenAPIDefinition(info = @Info(title = DEFINITION_TITLE, description = DEFINITION_DESCRIPTION, version = DEFINITION_VERSION))
